tarfiler_test.go,util_test.go: move utility functions for tests
